+++ Modules/Intent.Modules.Dapper/Templates/Repository/RepositoryTemplatePartial.cs
@@ -157,6 +157,7 @@ namespace Intent.Modules.Dapper.Templates.Repository
                     {
                         if (!method.Statements.Any())
                         {
+                            method.AddStatement($"// TODO: Implement {method.Name} ({@class.Name}) functionality");
                             method.AddStatement($"""throw new {UseType("System.NotImplementedException")}("Your implementation here...");""");
                         }
                     }