+++ Modules/Intent.Modules.Application.MediatR.FluentValidation/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.6.1
+
+- Improvement: Added `TODO` comments on `NotImplementedException`.
+
 ### Version 4.6.0
 
 - Improvement: Added Regular Expressions for Validation.